+++ reptor/api/NotesAPI.py
@@ -25,7 +25,7 @@ class NotesAPI(APIClient):
         if self.private_note:
             self.base_endpoint = urljoin(
                 self.reptor.get_config().get_server(),
-                f"api/v1/pentestusers/self/notes/",
+                "api/v1/pentestusers/self/notes/",
             )
         elif self.project_id:
             self.base_endpoint = urljoin(
@@ -38,7 +38,10 @@ class NotesAPI(APIClient):
         return self.reptor.get_config().get("private_note")
 
     def get_notes(self) -> typing.List[Note]:
-        """Gets list of notes"""
+        """
+        Gets list of notes
+        (Notes are not paginated)
+        """
         self.debug("Getting Notes List")
         response = self.get(self.base_endpoint)
         notes = list()